# [OTLP] Refactor ProtobufSerializer.WriteReservedLength() to improve performance