--- a/src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/Serializer/ProtobufSerializer.cs
+++ b/src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/Serializer/ProtobufSerializer.cs
@@ -72,35 +72,10 @@ internal static class ProtobufSerializer
         while (length > 0);
 
         var slice = buffer.AsSpan(writePosition, 4);
-
-        if (fourthByte.HasValue)
-        {
-            slice[0] = (byte)(firstByte.GetValueOrDefault() | 0x80);
-            slice[1] = (byte)(secondByte.GetValueOrDefault() | 0x80);
-            slice[2] = (byte)(thirdByte.GetValueOrDefault() | 0x80);
-            slice[3] = (byte)fourthByte.GetValueOrDefault();
-        }
-        else if (thirdByte.HasValue)
-        {
-            slice[0] = (byte)(firstByte.GetValueOrDefault() | 0x80);
-            slice[1] = (byte)(secondByte.GetValueOrDefault() | 0x80);
-            slice[2] = (byte)(thirdByte.GetValueOrDefault() | 0x80);
-            slice[3] = 0;
-        }
-        else if (secondByte.HasValue)
-        {
-            slice[0] = (byte)(firstByte.GetValueOrDefault() | 0x80);
-            slice[1] = (byte)(secondByte.GetValueOrDefault() | 0x80);
-            slice[2] = 0x80;
-            slice[3] = 0;
-        }
-        else
-        {
-            slice[0] = (byte)(firstByte.GetValueOrDefault() | 0x80);
-            slice[1] = 0x80;
-            slice[2] = 0x80;
-            slice[3] = 0;
-        }
+        slice[0] = (byte)(firstByte.GetValueOrDefault() | 0x80);
+        slice[1] = (byte)(secondByte.GetValueOrDefault() | 0x80);
+        slice[2] = (byte)(thirdByte.GetValueOrDefault() | 0x80);
+        slice[3] = (byte)fourthByte.GetValueOrDefault();
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]