@paulomorgado 2025-07-14
> > Other thoughts -utilize binary literals, instead of hex, for me it is easier to understand what happen under the hood.
> 
> I've done this for the instances of the two relevant values in the OTLP exporter.

Gotcha, @martincostello!

If you start messing with that, you'll end up at:

```csharp
internal static bool IncreaseBufferSize(ref byte[] buffer, OtlpSignalType otlpSignalType)
{
    if (buffer.Length >= MaxBufferSize)
    {
        OpenTelemetryProtocolExporterEventSource.Log.BufferExceededMaxSize(otlpSignalType.ToString(), buffer.Length);
        return false;
    }

    try
    {
        var newBufferSize = buffer.Length * 2;
        buffer = new byte[newBufferSize];
        return true;
    }
    catch (OutOfMemoryException)
    {
        OpenTelemetryProtocolExporterEventSource.Log.BufferResizeFailedDueToMemory(otlpSignalType.ToString());
        return false;
    }
}
```

This needs to change from the ground up and, to be honest, I can get the time to do the change, but not the time for fighting battles over code that should have never been written.