@paulomorgado 2025-07-17
@Kielek,

> > This needs to change from the ground up and, to be honest, I can get the time to do the change, but not the time for fighting battles over code that should have never been written.
> 
> Looking for your contributions, especially if it can significantly improve exporting time.

I'll have a look into it.