@rajkumar-rangaraj 2025-07-21
LGTM. Nice refactor! We made a decision to fix the length as 4 bytes for faster development. Later, we should refactor this to optimize buffer usage and encode the length using minimal varint size when appropriate.